Pi vision - Reader for blind using Raspberry Pi
The project is focused to help the blind people to read the sign boards so that it makes them independent. We make it possible by capturing the image from the pi camera. feed the image to image processing model that extracts the words. These words are then fed into an engine where it converts text to voice and outputs through speaker.